+++ changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Mocking of web request context with current calling context as parent so all BIFS work as expected
 - Updated all internal versions
 - Updated dependencies to latest versions
 - Updated all github actions